MG101
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@295 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Common/Manager/PackageLogReader.cpp
@@ -265,7 +265,7 @@ void MgPackageLogReader::ReadStatus()
     // Validate the current package status information.
     STRING packageName = m_statusInfo.GetPackageName();
     Ptr<MgDateTime> packageDate = m_statusInfo.GetPackageDate();
-    INT32 packageSize = m_statusInfo.GetPackageSize();
+    INT64 packageSize = m_statusInfo.GetPackageSize();
 
     if (packageName != actualPackageName 
      || packageSize != actualPackageSize
